Replace hub with gh and remove phantomjs
**Why**:
gh is the new and better official version of the GitHub CLI.

There are better tools for browser testing than PhantomJS these days,
like the `webdrivers`, `ferrum`, and `cuprite` gems. It's no longer
necessary to install a specific tool with Homebrew in order to run
feature/system tests in Ruby.